+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Development version 2.3.0
 
 **New Features**
+* Added support for headers in JSON-Batch requets, enabling JSON-Patch requests in JSON-Batch requests.
 
 
 ## Release version 2.2.0